Implement some missing variable functor evaluation routines
- ElemArg for finite elements
- ElemQpArg and ElemSideQpArg for finite volumes
