+++ framework/include/variables/MooseVariableField.h
@@ -345,7 +345,6 @@ public:
   virtual const DoFValue & nodalVectorTagValue(TagID tag) const = 0;
   virtual const DoFValue & vectorTagDofValue(TagID tag) const = 0;
 
-  virtual void meshChanged() override;
   virtual void residualSetup() override;
   virtual void jacobianSetup() override;
   virtual void timestepSetup() override;
@@ -361,95 +360,20 @@ public:
   bool hasBlocks(const SubdomainID id) const override { return BlockRestrictable::hasBlocks(id); }
 
 protected:
-  using FunctorArg = typename Moose::ADType<OutputType>::type;
-  using Moose::FunctorBase<FunctorArg>::evaluate;
-  using Moose::FunctorBase<FunctorArg>::evaluateGradient;
-  using Moose::FunctorBase<FunctorArg>::evaluateDot;
-  using typename Moose::FunctorBase<FunctorArg>::ValueType;
-  using typename Moose::FunctorBase<FunctorArg>::DotType;
-  using typename Moose::FunctorBase<FunctorArg>::GradientType;
-
-  using ElemQpArg = Moose::ElemQpArg;
-  using ElemSideQpArg = Moose::ElemSideQpArg;
-  using ElemPointArg = Moose::ElemPointArg;
-  using StateArg = Moose::StateArg;
-
   /**
    * Get the solution corresponding to the provided state
    */
-  const NumericVector<Number> & getSolution(const StateArg & state) const;
-
-  ValueType evaluate(const ElemQpArg & elem_qp, const StateArg & state) const override final;
-  ValueType evaluate(const ElemSideQpArg & elem_side_qp,
-                     const StateArg & state) const override final;
-  ValueType evaluate(const ElemPointArg & elem_point, const StateArg & state) const override final;
-
-  GradientType evaluateGradient(const ElemQpArg & elem_qp, const StateArg & state) const override;
-  GradientType evaluateGradient(const ElemSideQpArg & elem_side_qp,
-                                const StateArg & state) const override final;
-
-  DotType evaluateDot(const ElemQpArg & elem_qp, const StateArg & state) const override final;
-  DotType evaluateDot(const ElemSideQpArg & elem_side_qp,
-                      const StateArg & state) const override final;
+  const NumericVector<Number> & getSolution(const Moose::StateArg & state) const;
 
   /// the time integrator used for computing time derivatives
   const TimeIntegrator * const _time_integrator;
 
   /// A dummy ADReal variable
   mutable ADReal _ad_real_dummy = 0;
-
-private:
-  /**
-   * Compute the solution, gradient, and time derivative with provided shape functions
-   */
-  template <typename Shapes, typename Solution, typename GradShapes, typename GradSolution>
-  void computeSolution(const Elem * elem,
-                       const QBase *,
-                       const StateArg & state,
-                       const Shapes & phi,
-                       Solution & local_soln,
-                       const GradShapes & grad_phi,
-                       GradSolution & grad_local_soln,
-                       Solution & dot_local_soln) const;
-
-  /**
-   * Evaluate solution and gradient for the \p elem_qp argument
-   */
-  void evaluateOnElement(const ElemQpArg & elem_qp, const StateArg & state) const;
-
-  /**
-   * Evaluate solution and gradient for the \p elem_side_qp argument
-   */
-  void evaluateOnElementSide(const ElemSideQpArg & elem_side_qp, const StateArg & state) const;
-
-  /// Keep track of the current elem-qp functor element in order to enable local caching (e.g. if we
-  /// call evaluate on the same element, but just with a different quadrature point, we can return
-  /// previously computed results indexed at the different qp
-  mutable const Elem * _current_elem_qp_functor_elem = nullptr;
-
-  /// The values of the solution for the \p _current_elem_qp_functor_elem
-  mutable std::vector<ValueType> _current_elem_qp_functor_sln;
-
-  /// The values of the gradient for the \p _current_elem_qp_functor_elem
-  mutable std::vector<GradientType> _current_elem_qp_functor_gradient;
-
-  /// The values of the time derivative for the \p _current_elem_qp_functor_elem
-  mutable std::vector<DotType> _current_elem_qp_functor_dot;
-
-  /// Keep track of the current elem-side-qp functor element and side in order to enable local
-  /// caching (e.g. if we call evaluate with the same element and side, but just with a different
-  /// quadrature point, we can return previously computed results indexed at the different qp
-  mutable std::pair<const Elem *, unsigned int> _current_elem_side_qp_functor_elem_side{
-      nullptr, libMesh::invalid_uint};
-
-  /// The values of the solution for the \p _current_elem_side_qp_functor_elem_side
-  mutable std::vector<ValueType> _current_elem_side_qp_functor_sln;
-
-  /// The values of the gradient for the \p _current_elem_side_qp_functor_elem_side
-  mutable std::vector<GradientType> _current_elem_side_qp_functor_gradient;
-
-  /// The values of the time derivative for the \p _current_elem_side_qp_functor_elem_side
-  mutable std::vector<DotType> _current_elem_side_qp_functor_dot;
 };
 
-#define usingMooseVariableFieldMembers usingMooseVariableFieldBaseMembers
+#define usingMooseVariableFieldMembers                                                             \
+  usingMooseVariableFieldBaseMembers;                                                              \
+  using MooseVariableField<OutputType>::_time_integrator;                                          \
+  using MooseVariableField<OutputType>::_ad_real_dummy;                                            \
+  using MooseVariableField<OutputType>::getSolution
