+++ framework/src/variables/MooseVariableFE.C
@@ -14,6 +14,9 @@
 #include "DisplacedSystem.h"
 #include "Assembly.h"
 #include "MooseVariableData.h"
+#include "ArbitraryQuadrature.h"
+
+#include "libmesh/quadrature_monomial.h"
 
 template <>
 InputParameters
@@ -866,6 +869,365 @@ MooseVariableFE<OutputType>::evaluate(const NodeArg & node_arg, const StateArg &
   }
   else
     mooseError("RealEigenVector not yet supported for functors");
+}
+
+namespace
+{
+template <typename OutputType>
+struct FEBaseHelper
+{
+  typedef FEBase type;
+};
+
+template <>
+struct FEBaseHelper<RealVectorValue>
+{
+  typedef FEVectorBase type;
+};
+}
+
+template <typename OutputType>
+template <typename Shapes, typename Solution, typename GradShapes, typename GradSolution>
+void
+MooseVariableFE<OutputType>::computeSolution(const Elem * const elem,
+                                             const unsigned int n_qp,
+                                             const StateArg & state,
+                                             const Shapes & phi,
+                                             Solution & local_soln,
+                                             const GradShapes & grad_phi,
+                                             GradSolution & grad_local_soln,
+                                             Solution & dot_local_soln) const
+{
+  std::vector<dof_id_type> dof_indices;
+  _dof_map.dof_indices(elem, dof_indices, _var_num);
+  std::vector<ADReal> dof_values;
+  std::vector<ADReal> dof_values_dot;
+  dof_values.reserve(dof_indices.size());
+
+  const bool computing_dot = _time_integrator && _time_integrator->dt();
+  if (computing_dot)
+    dof_values_dot.reserve(dof_indices.size());
+
+  const bool do_derivatives = Moose::doDerivatives(_subproblem, _sys);
+  const auto & global_soln = getSolution(state);
+  for (const auto dof_index : dof_indices)
+  {
+    dof_values.push_back(ADReal(global_soln(dof_index)));
+    if (do_derivatives && state.state == 0)
+      Moose::derivInsert(dof_values.back().derivatives(), dof_index, 1.);
+    if (computing_dot)
+    {
+      if (_var_kind == Moose::VAR_NONLINEAR)
+      {
+        dof_values_dot.push_back(dof_values.back());
+        _time_integrator->computeADTimeDerivatives(
+            dof_values_dot.back(), dof_index, _ad_real_dummy);
+      }
+      else
+        dof_values_dot.push_back((*_sys.solutionUDot())(dof_index));
+    }
+  }
+
+  local_soln.resize(n_qp);
+  grad_local_soln.resize(n_qp);
+  if (computing_dot)
+    dot_local_soln.resize(n_qp);
+
+  for (const auto qp : make_range(n_qp))
+  {
+    local_soln[qp] = 0;
+    grad_local_soln[qp] = 0;
+    if (computing_dot)
+      dot_local_soln[qp] = 0;
+    for (const auto i : index_range(dof_indices))
+    {
+      local_soln[qp] += dof_values[i] * phi[i][qp];
+      grad_local_soln[qp] += dof_values[i] * grad_phi[i][qp];
+      if (computing_dot)
+        dot_local_soln[qp] += dof_values_dot[i] * phi[i][qp];
+    }
+  }
+}
+
+template <typename OutputType>
+void
+MooseVariableFE<OutputType>::evaluateOnElement(const ElemQpArg & elem_qp,
+                                               const StateArg & state,
+                                               const bool cache_eligible) const
+{
+  mooseAssert(this->hasBlocks(elem_qp.elem->subdomain_id()),
+              "This variable doesn't exist in the requested block!");
+
+  const Elem * const elem = elem_qp.elem;
+  if (!cache_eligible || (elem != _current_elem_qp_functor_elem))
+  {
+    const QBase * const qrule_template = elem_qp.qrule;
+
+    using FEBaseType = typename FEBaseHelper<OutputType>::type;
+    std::unique_ptr<FEBaseType> fe(FEBaseType::build(elem->dim(), _fe_type));
+    std::unique_ptr<QBase> qrule(QBase::build(
+        qrule_template->type(), qrule_template->get_dim(), qrule_template->get_order()));
+
+    const auto & phi = fe->get_phi();
+    const auto & dphi = fe->get_dphi();
+    fe->attach_quadrature_rule(qrule.get());
+    fe->reinit(elem);
+
+    computeSolution(elem,
+                    qrule->n_points(),
+                    state,
+                    phi,
+                    _current_elem_qp_functor_sln,
+                    dphi,
+                    _current_elem_qp_functor_gradient,
+                    _current_elem_qp_functor_dot);
+  }
+  if (cache_eligible)
+    _current_elem_qp_functor_elem = elem;
+  else
+    // These evaluations are not eligible for caching, e.g. maybe this is a single point quadrature
+    // rule evaluation at an arbitrary point and we don't want those evaluations to potentially be
+    // re-used when this function is called with a standard quadrature rule
+    _current_elem_qp_functor_elem = nullptr;
+}
+
+template <>
+void
+MooseVariableFE<RealEigenVector>::evaluateOnElement(const ElemQpArg &, const StateArg &, bool) const
+{
+  mooseError("evaluate not implemented for array variables");
+}
+
+template <typename OutputType>
+typename MooseVariableFE<OutputType>::ValueType
+MooseVariableFE<OutputType>::evaluate(const ElemQpArg & elem_qp, const StateArg & state) const
+{
+  evaluateOnElement(elem_qp, state, /*query_cache=*/true);
+  const auto qp = elem_qp.qp;
+  mooseAssert(qp < _current_elem_qp_functor_sln.size(),
+              "The requested " << qp << " is outside our solution size");
+  return _current_elem_qp_functor_sln[qp];
+}
+
+template <typename OutputType>
+typename MooseVariableFE<OutputType>::GradientType
+MooseVariableFE<OutputType>::evaluateGradient(const ElemQpArg & elem_qp,
+                                              const StateArg & state) const
+{
+  evaluateOnElement(elem_qp, state, /*query_cache=*/true);
+  const auto qp = elem_qp.qp;
+  mooseAssert(qp < _current_elem_qp_functor_gradient.size(),
+              "The requested " << qp << " is outside our gradient size");
+  return _current_elem_qp_functor_gradient[qp];
+}
+
+template <typename OutputType>
+typename MooseVariableFE<OutputType>::DotType
+MooseVariableFE<OutputType>::evaluateDot(const ElemQpArg & elem_qp, const StateArg & state) const
+{
+  mooseAssert(_time_integrator && _time_integrator->dt(),
+              "A time derivative is being requested but we do not have a time integrator so we'll "
+              "have no idea how to compute it");
+  evaluateOnElement(elem_qp, state, /*query_cache=*/true);
+  const auto qp = elem_qp.qp;
+  mooseAssert(qp < _current_elem_qp_functor_dot.size(),
+              "The requested " << qp << " is outside our dot size");
+  return _current_elem_qp_functor_dot[qp];
+}
+
+template <typename OutputType>
+typename MooseVariableFE<OutputType>::ValueType
+MooseVariableFE<OutputType>::evaluate(const ElemArg & elem_arg, const StateArg & state) const
+{
+  const QMonomial qrule(elem_arg.elem->dim(), CONSTANT);
+  // We can use whatever we want for the point argument since it won't be used
+  const ElemQpArg elem_qp_arg{elem_arg.elem, /*qp=*/0, &qrule, Point(0, 0, 0)};
+  evaluateOnElement(elem_qp_arg, state, /*cache_eligible=*/false);
+  const auto qp = elem_qp_arg.qp;
+  mooseAssert(qp < _current_elem_qp_functor_sln.size(),
+              "The requested " << qp << " is outside our solution size");
+  return _current_elem_qp_functor_sln[qp];
+}
+
+template <typename OutputType>
+typename MooseVariableFE<OutputType>::ValueType
+MooseVariableFE<OutputType>::evaluate(const ElemPointArg & elem_point_arg,
+                                      const StateArg & state) const
+{
+  mooseAssert(elem_point_arg.elem, "We need an Elem");
+  const Elem & elem = *elem_point_arg.elem;
+  const auto dim = elem.dim();
+  ArbitraryQuadrature qrule(dim);
+  const std::vector<Point> ref_point = {FEMap::inverse_map(dim, &elem, elem_point_arg.point)};
+  qrule.setPoints(ref_point);
+  // We can use whatever we want for the point argument since it won't be used
+  const ElemQpArg elem_qp_arg{elem_point_arg.elem, /*qp=*/0, &qrule, elem_point_arg.point};
+  evaluateOnElement(elem_qp_arg, state, /*cache_eligible=*/false);
+  const auto qp = elem_qp_arg.qp;
+  mooseAssert(qp < _current_elem_qp_functor_sln.size(),
+              "The requested " << qp << " is outside our solution size");
+  return _current_elem_qp_functor_sln[qp];
+}
+
+template <typename OutputType>
+void
+MooseVariableFE<OutputType>::evaluateOnElementSide(const ElemSideQpArg & elem_side_qp,
+                                                   const StateArg & state) const
+{
+  mooseAssert(this->hasBlocks(elem_side_qp.elem->subdomain_id()),
+              "This variable doesn't exist in the requested block!");
+
+  const Elem * const elem = elem_side_qp.elem;
+  const auto side = elem_side_qp.side;
+  if (elem != _current_elem_side_qp_functor_elem_side.first ||
+      side != _current_elem_side_qp_functor_elem_side.second)
+  {
+    _current_elem_side_qp_functor_elem_side = std::make_pair(elem, side);
+    const QBase * const qrule_template = elem_side_qp.qrule;
+
+    using FEBaseType = typename FEBaseHelper<OutputType>::type;
+    std::unique_ptr<FEBaseType> fe(FEBaseType::build(elem->dim(), _fe_type));
+    std::unique_ptr<QBase> qrule(QBase::build(
+        qrule_template->type(), qrule_template->get_dim(), qrule_template->get_order()));
+
+    const auto & phi = fe->get_phi();
+    const auto & dphi = fe->get_dphi();
+    fe->attach_quadrature_rule(qrule.get());
+    fe->reinit(elem, side);
+
+    computeSolution(elem,
+                    qrule->n_points(),
+                    state,
+                    phi,
+                    _current_elem_side_qp_functor_sln,
+                    dphi,
+                    _current_elem_side_qp_functor_gradient,
+                    _current_elem_side_qp_functor_dot);
+  }
+}
+
+template <>
+void
+MooseVariableFE<RealEigenVector>::evaluateOnElementSide(const ElemSideQpArg &,
+                                                        const StateArg &) const
+{
+  mooseError("evaluate not implemented for array variables");
+}
+
+template <typename OutputType>
+typename MooseVariableFE<OutputType>::ValueType
+MooseVariableFE<OutputType>::evaluate(const ElemSideQpArg & elem_side_qp,
+                                      const StateArg & state) const
+{
+  evaluateOnElementSide(elem_side_qp, state);
+  const auto qp = elem_side_qp.qp;
+  mooseAssert(qp < _current_elem_side_qp_functor_sln.size(),
+              "The requested " << qp << " is outside our solution size");
+  return _current_elem_side_qp_functor_sln[qp];
+}
+
+template <typename OutputType>
+typename MooseVariableFE<OutputType>::GradientType
+MooseVariableFE<OutputType>::evaluateGradient(const ElemSideQpArg & elem_side_qp,
+                                              const StateArg & state) const
+{
+  evaluateOnElementSide(elem_side_qp, state);
+  const auto qp = elem_side_qp.qp;
+  mooseAssert(qp < _current_elem_side_qp_functor_gradient.size(),
+              "The requested " << qp << " is outside our gradient size");
+  return _current_elem_side_qp_functor_gradient[qp];
+}
+
+template <typename OutputType>
+typename MooseVariableFE<OutputType>::DotType
+MooseVariableFE<OutputType>::evaluateDot(const ElemSideQpArg & elem_side_qp,
+                                         const StateArg & state) const
+{
+  mooseAssert(_time_integrator && _time_integrator->dt(),
+              "A time derivative is being requested but we do not have a time integrator so we'll "
+              "have no idea how to compute it");
+  evaluateOnElementSide(elem_side_qp, state);
+  const auto qp = elem_side_qp.qp;
+  mooseAssert(qp < _current_elem_side_qp_functor_dot.size(),
+              "The requested " << qp << " is outside our dot size");
+  return _current_elem_side_qp_functor_dot[qp];
+}
+
+template <>
+typename MooseVariableFE<RealEigenVector>::ValueType
+MooseVariableFE<RealEigenVector>::evaluate(const ElemQpArg &, const StateArg &) const
+{
+  mooseError(
+      "MooseVariableFE::evaluate(ElemQpArg &, const StateArg &) overload not implemented for "
+      "array variables");
+}
+
+template <>
+typename MooseVariableFE<RealEigenVector>::ValueType
+MooseVariableFE<RealEigenVector>::evaluate(const ElemSideQpArg &, const StateArg &) const
+{
+  mooseError("MooseVariableFE::evaluate(ElemSideQpArg &, const StateArg &) overload not "
+             "implemented for "
+             "array variables");
+}
+
+template <>
+typename MooseVariableFE<RealEigenVector>::GradientType
+MooseVariableFE<RealEigenVector>::evaluateGradient(const ElemQpArg &, const StateArg &) const
+{
+  mooseError("MooseVariableFE::evaluateGradient(ElemQpArg &, const StateArg &) overload not "
+             "implemented for array variables");
+}
+
+template <>
+typename MooseVariableFE<RealEigenVector>::GradientType
+MooseVariableFE<RealEigenVector>::evaluateGradient(const ElemSideQpArg &, const StateArg &) const
+{
+  mooseError("MooseVariableFE::evaluateGradient(ElemSideQpArg &, const StateArg &) overload not "
+             "implemented for array variables");
+}
+
+template <>
+typename MooseVariableFE<RealEigenVector>::DotType
+MooseVariableFE<RealEigenVector>::evaluateDot(const ElemQpArg &, const StateArg &) const
+{
+  mooseError("MooseVariableFE::evaluateDot(ElemQpArg &, const StateArg &) overload not "
+             "implemented for array variables");
+}
+
+template <>
+typename MooseVariableFE<RealEigenVector>::DotType
+MooseVariableFE<RealEigenVector>::evaluateDot(const ElemSideQpArg &, const StateArg &) const
+{
+  mooseError("MooseVariableFE::evaluateDot(ElemSideQpArg &, const StateArg &) overload not "
+             "implemented for array variables");
+}
+
+template <typename OutputType>
+void
+MooseVariableFE<OutputType>::meshChanged()
+{
+  _current_elem_qp_functor_elem = nullptr;
+  _current_elem_side_qp_functor_elem_side = std::make_pair(nullptr, libMesh::invalid_uint);
+  MooseVariableField<OutputType>::meshChanged();
+}
+
+template <typename OutputType>
+void
+MooseVariableFE<OutputType>::residualSetup()
+{
+  _current_elem_qp_functor_elem = nullptr;
+  _current_elem_side_qp_functor_elem_side = std::make_pair(nullptr, libMesh::invalid_uint);
+  MooseVariableField<OutputType>::residualSetup();
+}
+
+template <typename OutputType>
+void
+MooseVariableFE<OutputType>::jacobianSetup()
+{
+  _current_elem_qp_functor_elem = nullptr;
+  _current_elem_side_qp_functor_elem_side = std::make_pair(nullptr, libMesh::invalid_uint);
+  MooseVariableField<OutputType>::jacobianSetup();
 }
 
 template class MooseVariableFE<Real>;
